+++ core/src/kerr.c
@@ -2,7 +2,7 @@
 /******************************************************************************/
 /*                                                                            */
 /* RK0 - The Embedded Real-Time Kernel '0'                                    */
-/* VERSION: V0.19.3                                                           */
+/* VERSION: V0.20.0                                                           */
 /* You may obtain a copy of the License at :                                  */
 /* http://www.apache.org/licenses/LICENSE-2.0                                 */
 /*                                                                            */
@@ -66,6 +66,7 @@
     F(RK_FAULT_INVALID_OBJ)                                                    \
     F(RK_FAULT_INVALID_PARAM)                                                  \
     F(RK_FAULT_INVALID_TIMEOUT)                                                \
+    F(RK_FAULT_CHANNEL_NOT_ACTIVE)                                             \
     F(RK_FAULT_STACK_OVERFLOW)                                                 \
     F(RK_FAULT_TASK_COUNT_MISMATCH)                                            \
     F(RK_FAULT_KERNEL_VERSION)                                                 \